Fix file names to match class names
SA1649
