+++ src/Ocelot/Configuration/ServiceProviderConfiguration.cs
@@ -1,0 +1,14 @@
+﻿namespace Ocelot.Configuration
+{
+    public class ServiceProviderConfiguration
+    {
+        public ServiceProviderConfiguration(string serviceProviderHost, int serviceProviderPort)
+        {
+            ServiceProviderHost = serviceProviderHost;
+            ServiceProviderPort = serviceProviderPort;
+        }
+
+        public string ServiceProviderHost { get; private set; }
+        public int ServiceProviderPort { get; private set; }
+    }
+}
