--- src/Ocelot/DownstreamRouteFinder/UrlMatcher/UrlPathPlaceholderNameAndValue.cs
@@ -1,14 +1,0 @@
-namespace Ocelot.DownstreamRouteFinder.UrlMatcher
-{
-    public class PlaceholderNameAndValue
-    {
-        public PlaceholderNameAndValue(string name, string value)
-        {
-            Name = name;
-            Value = value;
-        }
-
-        public string Name {get;private set;}
-        public string Value {get;private set;}
-    }
-}
