--- src/Ocelot/Raft/ExcludeFromCoverage.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace Ocelot.Raft
-{
-    [AttributeUsage(AttributeTargets.Class|AttributeTargets.Method|AttributeTargets.Property)]
-    public class ExcludeFromCoverageAttribute : Attribute{}
-}
