+++ src/Ocelot/Raft/ExcludeFromCoverageAttribute.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace Ocelot.Raft
+{
+    [AttributeUsage(AttributeTargets.Class|AttributeTargets.Method|AttributeTargets.Property)]
+    public class ExcludeFromCoverageAttribute : Attribute{}
+}
