+++ src/Ocelot/Requester/ITracingHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace Ocelot.Requester
+{
+    public interface ITracingHandler
+    {
+    }
+}
