+++ src/Ocelot/Responses/ErrorResponseGeneric.cs
@@ -3,7 +3,9 @@ using Ocelot.Errors;
 
 namespace Ocelot.Responses
 {
+#pragma warning disable SA1649 // File name must match first type name
     public class ErrorResponse<T> : Response<T>
+#pragma warning restore SA1649 // File name must match first type name
     {
         public ErrorResponse(Error error) 
             : base(new List<Error> {error})
