+++ test/Ocelot.UnitTests/Wait.cs
@@ -1,0 +1,10 @@
+namespace Ocelot.UnitTests
+{
+    public class Wait
+    {
+        public static Waiter WaitFor(int milliSeconds)
+        {
+            return new Waiter(milliSeconds);
+        }
+    }
+}
